+++ app/models/DoubleAuctionMechanism.scala
@@ -135,7 +135,7 @@ case class DoubleAuctionMechanism(clearingMechanism: ActorRef, instrument: Asset
     * @param price price at which order is filled.
     * @param quantity quantity involved in the filled order.
     */
-  def generatePartialFill(ask: AskOrderLike, bid: BidOrderLike, price: Double, quantity: Int): Unit = {
+  def generatePartialFill(ask: AskOrderLike, bid: BidOrderLike, price: Double, quantity: Double): Unit = {
     val partialFill = PartialFill(ask.tradingPartyRef, bid.tradingPartyRef, instrument, price, quantity)
     log.info(s",${System.nanoTime()}" + partialFill.toString)
     updateReferencePrice(price)
@@ -149,7 +149,7 @@ case class DoubleAuctionMechanism(clearingMechanism: ActorRef, instrument: Asset
     * @param price price at which order is filled.
     * @param quantity quantity involved in the filled order.
     */
-  def generateTotalFill(ask: AskOrderLike, bid: BidOrderLike, price: Double, quantity: Int): Unit = {
+  def generateTotalFill(ask: AskOrderLike, bid: BidOrderLike, price: Double, quantity: Double): Unit = {
     val totalFill = TotalFill(ask.tradingPartyRef, bid.tradingPartyRef, instrument, price, quantity)
     log.info(s",${System.nanoTime()}" + totalFill.toString)
     updateReferencePrice(price)
